update `contactId` type
The type of `contactId` in `CommunicationLogEntity` and associated DAO methods has been changed from `Int` to `Long`. This aligns the `contactId` type with the primary key type of `ContactEntity` (which is `Long`).

+++ app/src/main/java/com/hollowvyn/kneatr/data/local/dao/CommunicationLogDao.kt
@@ -16,13 +16,13 @@ interface CommunicationLogDao {
     suspend fun insertCommunicationLog(communicationLog: CommunicationLogEntity)
 
     @Query("SELECT * FROM communication_logs WHERE contactId = :contactId")
-    fun getCommunicationLogsForContact(contactId: Int): Flow<List<CommunicationLogEntity>>
+    fun getCommunicationLogsForContact(contactId: Long): Flow<List<CommunicationLogEntity>>
 
     @Delete
     suspend fun deleteCommunicationLog(communicationLog: CommunicationLogEntity)
 
     @Query("DELETE FROM communication_logs WHERE contactId = :contactId")
-    suspend fun deleteCommunicationLogsForContact(contactId: Int)
+    suspend fun deleteCommunicationLogsForContact(contactId: Long)
 
     @Query("DELETE FROM communication_logs")
     suspend fun deleteAllCommunicationLogs()
